+++ Common/src/main/java/fuzs/slotcycler/client/handler/SlotRendererHandler.java
@@ -8,9 +8,9 @@ import fuzs.slotcycler.config.ClientConfig;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Gui;
 import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.core.NonNullList;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.HumanoidArm;
-import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.GameType;
@@ -21,88 +21,99 @@ public class SlotRendererHandler {
     public static void onHudRender(Gui gui, PoseStack matrixStack, float tickDelta, int screenWidth, int screenHeight) {
         Minecraft minecraft = Minecraft.getInstance();
         if (!minecraft.options.hideGui) {
-            RenderSystem.enableBlend();
-            RenderSystem.defaultBlendFunc();
-            RenderSystem.disableDepthTest();
             if (minecraft.gameMode.getPlayerMode() != GameType.SPECTATOR) {
-                if (SlotCycler.CONFIG.get(ClientConfig.class).slotsDisplayState == ClientConfig.SlotsDisplayState.NEVER) return;
-                renderAdditionalHotbar(minecraft, gui, matrixStack, tickDelta, screenWidth, screenHeight);
+                if (SlotCycler.CONFIG.get(ClientConfig.class).slotsDisplayState != ClientConfig.SlotsDisplayState.NEVER) {
+                    RenderSystem.enableBlend();
+                    RenderSystem.defaultBlendFunc();
+                    RenderSystem.disableDepthTest();
+                    renderAdditionalHotbar(minecraft, gui, matrixStack, tickDelta, screenWidth, screenHeight);
+                }
             }
         }
     }
 
-    private static void renderAdditionalHotbar(Minecraft minecraft, Gui gui, PoseStack poseStack, float partialTicks, int screenWidth, int screenHeight) {
-        if (SlotCycler.CONFIG.get(ClientConfig.class).slotsDisplayState == ClientConfig.SlotsDisplayState.KEY && KeyBindingHandler.cycleSlotsDisplay == 0) return;
+    private static void renderAdditionalHotbar(Minecraft minecraft, Gui gui, PoseStack poseStack, float partialTicks, final int screenWidth, final int screenHeight) {
+
+        if (KeyBindingHandler.getSlotsDisplayTicks() == 0 && SlotCycler.CONFIG.get(ClientConfig.class).slotsDisplayState == ClientConfig.SlotsDisplayState.KEY) {
+            return;
+        }
+
         Player player = getCameraPlayer(minecraft);
-        if (player != null) {
-            final int originalScreenHeight = screenHeight;
-            screenHeight -= SlotCycler.CONFIG.get(ClientConfig.class).slotsOffset;
-            // support Raised mod natively on Fabric
-            screenHeight -= ClientModServices.ABSTRACTIONS.getRaisedDistance();
-            if (SlotCycler.CONFIG.get(ClientConfig.class).slotsDisplayState == ClientConfig.SlotsDisplayState.KEY) {
-                screenHeight += (originalScreenHeight - screenHeight + 23) * (1.0F - Math.min(1.0F, (KeyBindingHandler.cycleSlotsDisplay - partialTicks) / 5.0F));
-            }
-            RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-            RenderSystem.setShader(GameRenderer::getPositionTexShader);
-            RenderSystem.setShaderTexture(0, WIDGETS_LOCATION);
-            final Inventory inventory = player.getInventory();
-            final int selected = inventory.selected;
-            ItemStack mainStack = inventory.items.get(selected);
-            HumanoidArm humanoidarm = player.getMainArm().getOpposite();
-            int leftSlot = getRightSlot(inventory, selected);
-            int rightSlot = getLeftSlot(inventory, selected);
-            if (humanoidarm == HumanoidArm.LEFT) {
-                if (leftSlot == -1) {
-                    return;
-                }
-            } else if (rightSlot == -1) {
+        if (player == null) return;
+
+        int posY = screenHeight;
+        posY -= SlotCycler.CONFIG.get(ClientConfig.class).slotsYOffset;
+        // support Raised mod natively on Fabric
+        posY -= ClientModServices.ABSTRACTIONS.getRaisedDistance();
+        if (SlotCycler.CONFIG.get(ClientConfig.class).slotsDisplayState == ClientConfig.SlotsDisplayState.KEY) {
+            posY += (screenHeight - posY + 23) * (1.0F - Math.min(1.0F, (KeyBindingHandler.getSlotsDisplayTicks() - partialTicks) / 5.0F));
+        }
+
+        NonNullList<ItemStack> items = player.getInventory().items;
+        int selected = player.getInventory().selected;
+        int leftSlot = getRightSlot(items, selected);
+        int rightSlot = getLeftSlot(items, selected);
+        boolean renderToRight = player.getMainArm() == HumanoidArm.RIGHT;
+        if (renderToRight) {
+            if (leftSlot == -1) {
                 return;
             }
-            ItemStack leftStack = inventory.items.get(leftSlot);
-            ItemStack rightStack = inventory.items.get(rightSlot);
-            if (humanoidarm == HumanoidArm.LEFT) {
-                if (leftSlot <= rightSlot) {
-                    rightStack = ItemStack.EMPTY;
-                }
-            } else {
-                if (leftSlot <= rightSlot) {
-                    leftStack = ItemStack.EMPTY;
-                }
+        } else if (rightSlot == -1) {
+            return;
+        }
+
+        ItemStack mainStack = items.get(selected);
+        ItemStack leftStack = items.get(leftSlot);
+        ItemStack rightStack = items.get(rightSlot);
+        if (renderToRight) {
+            if (leftSlot <= rightSlot) {
+                rightStack = ItemStack.EMPTY;
             }
-            final int screenCenter = screenWidth / 2;
-            if (humanoidarm == HumanoidArm.LEFT) {
-                gui.blit(poseStack, screenCenter + 91, screenHeight - 23, 53, 22, 29, 24);
-                if (!rightStack.isEmpty()) {
-                    gui.blit(poseStack, screenCenter + 91 + 40, screenHeight - 23, 53, 22, 29, 24);
-                }
-                gui.blit(poseStack, screenCenter + 91 + 28, screenHeight - 22, 21, 0, 20, 22);
-                gui.blit(poseStack, screenCenter + 91 + 26, screenHeight - 22 - 1, 0, 22, 24, 24);
-            } else {
-                if (!leftStack.isEmpty()) {
-                    gui.blit(poseStack, screenCenter - 91 - 29 - 40, screenHeight - 23, 24, 22, 29, 24);
-                }
-                gui.blit(poseStack, screenCenter - 91 - 29, screenHeight - 23, 24, 22, 29, 24);
-                gui.blit(poseStack, screenCenter - 91 - 29 - 19, screenHeight - 22, 21, 0, 20, 22);
-                gui.blit(poseStack, screenCenter - 91 - 29 - 21, screenHeight - 22 - 1, 0, 22, 24, 24);
+        } else {
+            if (leftSlot <= rightSlot) {
+                leftStack = ItemStack.EMPTY;
             }
-            int j2 = screenHeight - 16 - 3;
-            if (humanoidarm == HumanoidArm.LEFT) {
-                renderSlot(minecraft, screenCenter + 91 + 10, j2, partialTicks, player, leftStack);
-                renderSlot(minecraft, screenCenter + 91 + 10 + 20, j2, partialTicks, player, mainStack);
-                renderSlot(minecraft, screenCenter + 91 + 10 + 20 + 20, j2, partialTicks, player, rightStack);
-            } else {
-                renderSlot(minecraft, screenCenter - 91 - 26, j2, partialTicks, player, rightStack);
-                renderSlot(minecraft, screenCenter - 91 - 26 - 20, j2, partialTicks, player, mainStack);
-                renderSlot(minecraft, screenCenter - 91 - 26 - 20 - 20, j2, partialTicks, player, leftStack);
+        }
+
+        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
+        RenderSystem.setShader(GameRenderer::getPositionTexShader);
+        RenderSystem.setShaderTexture(0, WIDGETS_LOCATION);
+        
+        int posX = screenWidth / 2;
+        posX += (91 + SlotCycler.CONFIG.get(ClientConfig.class).slotsXOffset) * (renderToRight ? 1 : -1);
+        if (renderToRight) {
+            gui.blit(poseStack, posX, posY - 23, 53, 22, 29, 24);
+            if (!rightStack.isEmpty()) {
+                gui.blit(poseStack, posX + 40, posY - 23, 53, 22, 29, 24);
             }
+            gui.blit(poseStack, posX + 28, posY - 22, 21, 0, 20, 22);
+            gui.blit(poseStack, posX + 26, posY - 22 - 1, 0, 22, 24, 24);
+        } else {
+            if (!leftStack.isEmpty()) {
+                gui.blit(poseStack, posX - 29 - 40, posY - 23, 24, 22, 29, 24);
+            }
+            gui.blit(poseStack, posX - 29, posY - 23, 24, 22, 29, 24);
+            gui.blit(poseStack, posX - 29 - 19, posY - 22, 21, 0, 20, 22);
+            gui.blit(poseStack, posX - 29 - 21, posY - 22 - 1, 0, 22, 24, 24);
+        }
+
+        posY -= 16 + 3;
+        if (renderToRight) {
+            renderItemInSlot(minecraft, posX + 10, posY, partialTicks, player, leftStack);
+            renderItemInSlot(minecraft, posX + 10 + 20, posY, partialTicks, player, mainStack);
+            renderItemInSlot(minecraft, posX + 10 + 20 + 20, posY, partialTicks, player, rightStack);
+        } else {
+            renderItemInSlot(minecraft, posX - 26, posY, partialTicks, player, rightStack);
+            renderItemInSlot(minecraft, posX - 26 - 20, posY, partialTicks, player, mainStack);
+            renderItemInSlot(minecraft, posX - 26 - 20 - 20, posY, partialTicks, player, leftStack);
         }
     }
 
-    private static int getLeftSlot(Inventory inventory, int selected) {
+    private static int getLeftSlot(NonNullList<ItemStack> items, int selected) {
         int leftSlot = -1;
         for (int i = 1; i < 4; i++) {
             int slotIndex = i * 9 + selected;
-            if (!inventory.items.get(slotIndex).isEmpty()) {
+            if (!items.get(slotIndex).isEmpty()) {
                 leftSlot = slotIndex;
                 break;
             }
@@ -110,11 +121,11 @@ public class SlotRendererHandler {
         return leftSlot;
     }
 
-    private static int getRightSlot(Inventory inventory, int selected) {
+    private static int getRightSlot(NonNullList<ItemStack> items, int selected) {
         int rightSlot = -1;
         for (int i = 3; i > 0; i--) {
             int slotIndex = i * 9 + selected;
-            if (!inventory.items.get(slotIndex).isEmpty()) {
+            if (!items.get(slotIndex).isEmpty()) {
                 rightSlot = slotIndex;
                 break;
             }
@@ -126,25 +137,24 @@ public class SlotRendererHandler {
         return !(minecraft.getCameraEntity() instanceof Player) ? null : (Player) minecraft.getCameraEntity();
     }
 
-    private static void renderSlot(Minecraft minecraft, int p_168678_, int p_168679_, float p_168680_, Player p_168681_, ItemStack p_168682_) {
-        if (!p_168682_.isEmpty()) {
-            PoseStack posestack = RenderSystem.getModelViewStack();
-            float f = (float)p_168682_.getPopTime() - p_168680_;
-            if (f > 0.0F) {
-                float f1 = 1.0F + f / 5.0F;
-                posestack.pushPose();
-                posestack.translate((double)(p_168678_ + 8), (double)(p_168679_ + 12), 0.0D);
-                posestack.scale(1.0F / f1, (f1 + 1.0F) / 2.0F, 1.0F);
-                posestack.translate((double)(-(p_168678_ + 8)), (double)(-(p_168679_ + 12)), 0.0D);
-                RenderSystem.applyModelViewMatrix();
-            }
-            minecraft.getItemRenderer().renderAndDecorateItem(p_168681_, p_168682_, p_168678_, p_168679_, 0);
-            RenderSystem.setShader(GameRenderer::getPositionColorShader);
-            if (f > 0.0F) {
-                posestack.popPose();
-                RenderSystem.applyModelViewMatrix();
-            }
-            minecraft.getItemRenderer().renderGuiItemDecorations(minecraft.font, p_168682_, p_168678_, p_168679_);
+    private static void renderItemInSlot(Minecraft minecraft, int posX, int posY, float partialTicks, Player player, ItemStack stack) {
+        if (stack.isEmpty()) return;
+        PoseStack posestack = RenderSystem.getModelViewStack();
+        float f = (float) stack.getPopTime() - partialTicks;
+        if (f > 0.0F) {
+            float f1 = 1.0F + f / 5.0F;
+            posestack.pushPose();
+            posestack.translate(posX + 8, posY + 12, 0.0D);
+            posestack.scale(1.0F / f1, (f1 + 1.0F) / 2.0F, 1.0F);
+            posestack.translate(-(posX + 8), -(posY + 12), 0.0D);
+            RenderSystem.applyModelViewMatrix();
         }
+        minecraft.getItemRenderer().renderAndDecorateItem(player, stack, posX, posY, 0);
+        RenderSystem.setShader(GameRenderer::getPositionColorShader);
+        if (f > 0.0F) {
+            posestack.popPose();
+            RenderSystem.applyModelViewMatrix();
+        }
+        minecraft.getItemRenderer().renderGuiItemDecorations(minecraft.font, stack, posX, posY);
     }
 }