+++ Fabric/src/main/java/fuzs/slotcycler/mixin/client/MouseHandlerFabricMixin.java
@@ -12,7 +12,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(MouseHandler.class)
-abstract class MouseHandlerMixin {
+abstract class MouseHandlerFabricMixin {
 	@Shadow
 	@Final
 	private Minecraft minecraft;